Update README.md
This project aims to develop a text analytics tool using R for the NYU MASY program. With the deeper usage of the syllabus for both students and the professors, the content of the syllabus affects the recognition of the course, which will influence the course selection of the students. In this way, this project study aims at creating computer-based tool that matches a syllabus to jobs via keywords to discover what occupations the course prepares students for and analyzes gaps in knowledge and skills covered by the course. In a word, this project provides the solution for professors by giving the results keywords frequency to help modify their syllabus. In this project study, basic techniques like similarity scoring, term frequency analysis, descriptive statistics, and different charts are covered. Besides, this project study will display the project details, including the company, sponsors, evaluation, opportunity, and the challenges or risks of this project.